requirements.txt
fastapi
uvicorn
azure-cognitiveservices-speech
yt-dlp
requests
python-dotenv